More cleanups.
